requirements/common.txt
Flask==1.1.1
flask-restplus==0.13.0
beautifulsoup4==4.8.0
python-slugify==3.0.4
celery==4.3.0
redis==3.3.8
requests==2.22.0